% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwbs.R
\name{dwbs}
\alias{dwbs}
\title{Weibull Birnbaum-Saunders Density Function}
\usage{
dwbs(t, alpha, beta, a, b)
}
\arguments{
\item{t}{Time or failure time (t > 0)}

\item{alpha}{Shape parameter (alpha > 0)}

\item{beta}{Scale parameter (beta > 0)}

\item{a}{Shape parameter (a > 0)}

\item{b}{Shape parameter (b > 0)}
}
\value{
Density value of Weibull-Birnbaum-Saunders distribution at t
}
\description{
Weibull Birnbaum-Saunders Density Function
}
\examples{
dwbs(2.5,alpha=1.5,beta=3,a=2,b=5)
}
