% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failure.R
\name{failure}
\alias{failure}
\title{Calculate the WBS Hazard Rate Function}
\usage{
failure(t, alpha, beta, a, b)
}
\arguments{
\item{t}{A vector of positive quantiles (time).}

\item{alpha}{The shape parameter 'alpha' (must be positive).}

\item{beta}{The scale parameter 'beta' (must be positive).}

\item{a}{The shape parameter 'a' (must be positive).}

\item{b}{The shape parameter 'b' (must be positive).}
}
\value{
A vector of hazard rates, h(t).
}
\description{
Computes the hazard rate function (failure rate) h(t) for the
Weibull Birnbaum-Saunders (WBS) distribution.
}
\examples{
failure(t = 2.5, alpha = 1.5, beta = 0.1, a = 0.75, b = 0.25)
}
