% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kswbs.R
\name{kswbs}
\alias{kswbs}
\title{Kolmogorov-Smirnov Test for Weibull Birnbaum-Saunders Distribution}
\usage{
kswbs(
  y,
  alpha,
  beta,
  a,
  b,
  alternative = c("two.sided", "less", "greater"),
  plot = FALSE
)
}
\arguments{
\item{y}{A numeric vector of data points.}

\item{alpha}{Shape parameter (alpha > 0) of the distribution.}

\item{beta}{Scale parameter (beta > 0) of the distribution.}

\item{a}{Shape parameter (a > 0) of the distribution.}

\item{b}{Shape parameter (b > 0) of the distribution.}

\item{alternative}{One of "two.sided" (default), "less", or "greater".}

\item{plot}{Logical; if TRUE, plots the CDFs.}
}
\value{
A list with statistic and p-value.
}
\description{
Kolmogorov-Smirnov Test for Weibull Birnbaum-Saunders Distribution
}
\examples{
alpha=1.5; beta=0.1; a=0.75;b=0.25
set.seed(152)
y <- rwbs(100,alpha,beta,a,b)
kswbs(y,alpha,beta,a,b,alternative = "two.sided", plot=FALSE)
}
