% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelwbs.R
\name{modelwbs}
\alias{modelwbs}
\title{Maximum Likelihood Estimation for Weibull Birnbaum-Saunders Distribution}
\usage{
modelwbs(y, plot = "FALSE")
}
\arguments{
\item{y}{A numeric vector of observed data.}

\item{plot}{Logical; if TRUE, plots the Q-Q plot of residuals.}
}
\value{
A list containing the estimated parameters, AIC, BIC, and a Q-Q plot of residuals.
}
\description{
Maximum Likelihood Estimation for Weibull Birnbaum-Saunders Distribution
}
\examples{

alpha=1.5; beta=0.1; a=0.75;b=0.25
set.seed(123)
y=rwbs(100,alpha,beta,a,b)
modelwbs(y,plot='FALSE')
}
