% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwbs.R
\name{pwbs}
\alias{pwbs}
\title{Weibull Birnbaum-Saunders Cumulative Distribution Function}
\usage{
pwbs(t, alpha, beta, a, b)
}
\arguments{
\item{t}{Time or failure time (t > 0)}

\item{alpha}{Shape parameter (alpha > 0)}

\item{beta}{Scale parameter (beta > 0)}

\item{a}{Shape parameter (a > 0)}

\item{b}{Shape parameter (b > 0)}
}
\value{
CDF value of Weibull Birnbaum-Saunders distribution at t
}
\description{
Weibull Birnbaum-Saunders Cumulative Distribution Function
}
\examples{
pwbs(0.5,alpha=1.5,beta=2,a=0.5,b=1)
}
