% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwbs.R
\name{qwbs}
\alias{qwbs}
\title{Weibull Birnbaum-Saunders Quantile Function}
\usage{
qwbs(u, alpha, beta, a, b)
}
\arguments{
\item{u}{Probability value (u in [0, 1])}

\item{alpha}{Shape parameter (alpha > 0)}

\item{beta}{Scale parameter (beta > 0)}

\item{a}{Shape parameter (a > 0)}

\item{b}{Shape parameter (b > 0)}
}
\value{
Quantile corresponding to probability u
}
\description{
Weibull Birnbaum-Saunders Quantile Function
}
\examples{
qwbs(0.5, 0.5, 1, 1, 2)
}
