% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Calculate the WBS Reliability Function}
\usage{
reliability(t, alpha, beta, a, b)
}
\arguments{
\item{t}{A vector of positive quantiles (time).}

\item{alpha}{The shape parameter 'alpha' (must be positive).}

\item{beta}{The scale parameter 'beta' (must be positive).}

\item{a}{The shape parameter 'a' (must be positive).}

\item{b}{The shape parameter 'b' (must be positive).}
}
\value{
A vector of probabilities, R(t) = P(T > t).
}
\description{
Computes the reliability function (survival function) R(t) for the
Weibull Birnbaum-Saunders (WBS) distribution.
}
\examples{
reliability(t = 2.5, alpha = 1.5, beta = 0.1, a = 0.75, b = 0.25)
}
