% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwbs.R
\name{rwbs}
\alias{rwbs}
\title{Random Number Generation for Weibull Birnbaum-Saunders Distribution}
\usage{
rwbs(n, alpha, beta, a, b)
}
\arguments{
\item{n}{Number of random values to generate}

\item{alpha}{Shape parameter (alpha > 0)}

\item{beta}{Scale parameter (beta > 0)}

\item{a}{Shape parameter (a > 0)}

\item{b}{Shape parameter (b > 0)}
}
\value{
A vector of n random values following the WBS distribution
}
\description{
Random Number Generation for Weibull Birnbaum-Saunders Distribution
}
\examples{
rwbs(10, 0.5, 1, 1, 2)

}
