% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sbs}
\alias{plot.sbs}
\title{Plot for an 'sbs' object}
\usage{
\method{plot}{sbs}(x, cpt, ...)
}
\arguments{
\item{x}{an object of class 'sbs', returned by \code{\link{sbs}}}

\item{cpt}{a vector of integers with localisations of change-points}

\item{...}{other parameters which may be passed to \code{plot} and \code{changepoints}}
}
\description{
Plots the input vector used to generate 'sbs' object \code{x} with fitted piecewise constant function, equal to the mean
between change-points specified in \code{cpt}.
}
\details{
When \code{cpt} is omitted, the function automatically finds change-points 
using \code{changepoints} function with a default value of the threshold.
}
\seealso{
\code{\link{sbs}}  \code{\link{changepoints}}
}
