% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.wbs}
\alias{plot.wbs}
\title{Plot for a 'wbs' object}
\usage{
\method{plot}{wbs}(x, cpt, ...)
}
\arguments{
\item{x}{an object of class 'wbs', returned by \code{\link{wbs}}}

\item{cpt}{a vector of integers with localisations of change-points}

\item{...}{other parameters which may be passed to \code{plot} and \code{changepoints}}
}
\description{
Plots the input vector used to generate 'wbs' object \code{x} with fitted piecewise constant function, equal to the mean
between change-points specified in \code{cpt}.
}
\details{
When \code{cpt} is omitted, the function automatically finds change-points 
using \code{changepoints} function with strengthened Schwarz Information Criterion as a stopping criterion for the WBS algorithm.
}
\seealso{
\code{\link{wbs}}  \code{\link{changepoints}} \code{\link{ssic.penalty}}
}
