% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sbs}
\alias{print.sbs}
\title{Print for an 'sbs' object}
\usage{
\method{print}{sbs}(x, ...)
}
\arguments{
\item{x}{an object of class 'sbs'}

\item{...}{further arguments passed to \code{print} method}
}
\description{
Print for an 'sbs' object
}
\seealso{
\code{\link{sbs}}
}
