% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.wbs}
\alias{print.wbs}
\title{Print for a 'wbs' object}
\usage{
\method{print}{wbs}(x, ...)
}
\arguments{
\item{x}{an object of class 'wbs'}

\item{...}{further arguments passed to \code{print} method}
}
\description{
Print for a 'wbs' object
}
\seealso{
\code{\link{wbs}}
}
