% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.intervals.R
\name{random.intervals}
\alias{random.intervals}
\title{Random intervals}
\usage{
random.intervals(n, M)
}
\arguments{
\item{n}{a number of endpoints to choose from}

\item{M}{a number of intervals to generate}
}
\value{
a \code{M} by 2 matrix with start (first column) and end (second column) points of an interval in each row
}
\description{
The function generates \code{M} intervals, whose endpoints are 
are drawn uniformly without replacements from \code{1},\code{2},..., \code{n}. This routine can be
used inside \code{\link{wbs}} function and is typically not called directly by the user.
}
\examples{
random.intervals(10,100)
}
\seealso{
\code{\link{fixed.intervals}} \code{\link{wbs}}
}
